# Performance Optimization & Enhanced Template Rendering