… PERFORMANCE.md
… src/render.js
… test/performance/README.md
… test/performance/profile-bottlenecks.js